Help channels: move message functions to message module
